fixed bug related to hbmv

+++ lib/mkl/libonemkl.jl
@@ -269,7 +269,7 @@ end
 
 function onemklZhbmv(device_queue, uplo, n, k, alpha, a, lda, x, incx, beta, y, incy)
     @ccall liboneapi_support.onemklZhbmv(device_queue::syclQueue_t, uplo::onemklUplo,
-                                         n::Int64, k::Int64, alpha::ComplexF32, a::ZePtr{ComplexF64},
+                                         n::Int64, k::Int64, alpha::ComplexF64, a::ZePtr{ComplexF64},
                                          lda::Int64, x::ZePtr{ComplexF64}, incx::Int64, beta::ComplexF64,
                                          y::ZePtr{ComplexF64}, incy::Int64)::Cvoid
 end
